First attempt at a better @configure specification
Implemented a 'spec' argument to @configure which will handle the uuid,
input entity and transform set specification for each type of input you
want to handle.
This merges the three separate arguments to @configure which was kind of
clumsy to define and was an easy source of error if you either of the
three arguments didn't match in length, etc.

The new 'spec' is documented in transform.plate, but the basic idea is
to combine the arguments in a list of triples (uuid, input entity and
list of transform sets).

Backwards compatibility is maintained, by populating the 'old' fields,
thus the original code for generating the local transforms will still
work with this new style of specification.

Also added some more fields, that will be used in the future and
generally updated the structure and comments in the template files.